requirements.txt
# requirements.txt

# VSCode should detect this upon making a virtual environment, 
# then loading the root folder into it's explorer. 

# If it doesn't, please navigate to the folder's destination your terminal. 
# Then, run the following:

# pip install -r requirements.txt 

# Core data handling 
pandas>=2.1.3
numpy>=1.26.2

# Data formatting and storage
duckdb>=1.1.2
minio>=7.2.14
pyarrow>=18.1.0  # Required for parquet support
fastparquet>=2024.11.0
dbt-adapters>=1.13.2
dbt-common>=1.14.0
dbt-core>=1.9.1
dbt-extractor>=0.5.1
dbt-semantic-interfaces>=0.7.4

# Data generation 
Faker>=25.3.0
user-agents>=2.2.0

# Logging 
python-dotenv>=1.0.1
pathlib>=1.0.1

# Date & time handling
python-dateutil>=2.8.2
datetime>=5.4

# Optional: Development dependencies
# pytest>=7.0.0  # For testing
# black>=22.0.0  # For code formatting
